+++ src/controllers/masjidController.js
@@ -3,6 +3,10 @@ const MasjidService = require("../services/masjidService");
 class MasjidController {
   static async create(req, res) {
     try {
+      const user = req.user;
+      if (user.role !== "author" && user.role !== "admin") {
+        return res.status(403).json({ message: "Forbidden" });
+      }
       const masjid = await MasjidService.createMasjid(req.body);
       res.status(201).json(masjid);
     } catch (error) {
@@ -33,6 +37,10 @@ class MasjidController {
 
   static async update(req, res) {
     try {
+      const user = req.user;
+      if (user.role !== "admin") {
+        return res.status(403).json({ message: "Forbidden" });
+      }
       const masjid = await MasjidService.updateMasjid(req.params.id, req.body);
       if (!masjid) {
         return res.status(404).json({ message: "Masjid not found" });
@@ -45,6 +53,10 @@ class MasjidController {
 
   static async delete(req, res) {
     try {
+      const user = req.user;
+      if (user.role !== "admin") {
+        return res.status(403).json({ message: "Forbidden" });
+      }
       const masjid = await MasjidService.deleteMasjid(req.params.id);
       if (!masjid) {
         return res.status(404).json({ message: "Masjid not found" });
